refactor(mission-control): split FlightPlanBlueprint.tsx into subcomponents
Extracts 1920-line FlightPlanBlueprint.tsx into focused modules:

- BlueprintLayout.ts      — computeLayout(), INTEGRATION_LABELS, layout constants
- BlueprintReport.ts      — exportFullReport(), shortenClusterName(), CSS tokens
- BlueprintInfoPanels.tsx — ProjectInfoPanel, ClusterInfoPanel, DeployModeInfoPanel,
                            GaugeRow, generateDefaultPhases, getDependencyNotes

FlightPlanBlueprint.tsx is reduced to ~890 lines (the main component only).
All imports in MissionControlDialog.tsx remain unchanged (lazy import preserved).

Fixes #8624
